+++ cmdline/standalone_cli_tools/common/CmdLineArgs.cpp
@@ -63,8 +63,8 @@ static struct option verifyFileLongOption[] = {
 
 static struct option changeStorageClassLongOption[] = {
   {"id", required_argument, nullptr, 'I'},
-  {"filename", required_argument, nullptr, 'F'},
-  {"storage.class.name", required_argument, nullptr, 'n'},
+  {"json", required_argument, nullptr, 'j'},
+  {"storageclassname", required_argument, nullptr, 'n'},
   {"frequenzy", required_argument, nullptr, 't'},
   {"help", no_argument, nullptr, 'h'},
   {nullptr, 0, nullptr, 0}
@@ -88,7 +88,7 @@ std::map<StandaloneCliTool, const char*> shortopts = {
   {StandaloneCliTool::RESTORE_FILES, "I:i:f:F:v:c:hd:"},
   {StandaloneCliTool::CTA_SEND_EVENT, "i:e:u:g:"},
   {StandaloneCliTool::CTA_VERIFY_FILE, "I:F:i:u:g:v:h:"},
-  {StandaloneCliTool::CTA_CHANGE_STORAGE_CLASS, "I:F:n:t:h:"},
+  {StandaloneCliTool::CTA_CHANGE_STORAGE_CLASS, "I:j:n:t:h:"},
   {StandaloneCliTool::EOS_NAMESPACE_INJECTION, "j:h:"},
 };
 
@@ -270,7 +270,7 @@ void CmdLineArgs::printUsage(std::ostream &os) const {
     break;
   case StandaloneCliTool::CTA_CHANGE_STORAGE_CLASS :
     os << "    Usage:" << std::endl <<
-    "    cta-change-storage-class --id/-I <archiveFileID> | --filename/-F <filename> --storage.class.name/-n <storageClassName> [--frequenzy/-t <eosRequestFrequency>]" << std::endl << std::endl;
+    "    cta-change-storage-class --id/-I <archiveFileID> | --json/-j <path> --storageclassname/-n <storageClassName> [--frequenzy/-t <eosRequestFrequency>]" << std::endl << std::endl;
     break;
   case StandaloneCliTool::EOS_NAMESPACE_INJECTION :
     os << "    Usage:" << std::endl <<