+++ ReleaseNotes.md
@@ -11,6 +11,7 @@
 - cta/CTA#248 - Clean up output from cta-readtp
 - cta/CTA#251 - Increase free drive STALE threshold to 4 hours
 - cta/CTA#218 - Do not retry during repack requests
+- cta/CTA#252 - Update cta-change-storage-class to accept json file as input
 ### Bug Fixes
 - cta/CTA#181 - cta-statistics-update can fail for catalogues in postgres
 - cta/CTA#189 - Avoid postgres logging frequent warnings about no transaction in progress